requirements.txt
opencv-python
mediapipe
numpy
matplotlib
pandas
sklearn
joblib
lightgbm
optuna
PyYAML
shap